增加$.dropdownlist宽度补偿参数w，用于补偿input-group样式导致的下拉列右边框宽度为0时，选项框宽度显示不对齐的情况

+++ oui/form/oui.form.js
@@ -1894,7 +1894,7 @@
                 }
                 var pos = $.setPanelPosition(elem, elem.optbox, cfg);
                 if (cfg.x) {
-                    elem.optbox.style.left = parseInt(elem.optbox.style.left, 10) + cfg.x + 'px';
+                    elem.optbox.style.left = parseFloat(elem.optbox.style.left, 10) + cfg.x + 'px';
                 }
                 if (action) {
                     $.addClass(elem.optbox, 'input-opt-panel-box-' + pos.position);
